+++ x/btcstaking/keeper/genesis.go
@@ -2,7 +2,6 @@ package keeper
 
 import (
 	"context"
-	"encoding/hex"
 	"fmt"
 	"math"
 	"sort"
@@ -63,34 +62,12 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 		}
 	}
 
-	for _, hStr := range gs.AllowedStakingTxHashes {
-		// hashes are hex encoded for better readability
-		bz, err := hex.DecodeString(hStr)
-		if err != nil {
-			return fmt.Errorf("error decoding tx hash: %w", err)
-		}
-		if err := k.AllowedStakingTxHashesKeySet.Set(ctx, bz); err != nil {
-			return err
-		}
-	}
-
 	for _, fpAddrStr := range gs.FpBbnAddr {
 		fpAddr, err := sdk.AccAddressFromBech32(fpAddrStr)
 		if err != nil {
 			return fmt.Errorf("error decoding fp addr %s: %w", fpAddrStr, err)
 		}
 		if err := k.SetFpBbnAddr(ctx, fpAddr); err != nil {
-			return err
-		}
-	}
-
-	for _, hStr := range gs.AllowedMultiStakingTxHashes {
-		// hashes are hex encoded for better readability
-		bz, err := hex.DecodeString(hStr)
-		if err != nil {
-			return fmt.Errorf("error decoding tx hash: %w", err)
-		}
-		if err := k.allowedMultiStakingTxHashesKeySet.Set(ctx, bz); err != nil {
 			return err
 		}
 	}
@@ -140,17 +117,7 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
-	txHashes, err := k.allowedStakingTxHashes(ctx)
-	if err != nil {
-		return nil, err
-	}
-
 	fpBbnAddr, err := k.fpBtcPkByAddrEntries(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	multiStakingTxHashes, err := k.allowedMultiStakingTxHashes(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -161,18 +128,16 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 	}
 
 	return &types.GenesisState{
-		Params:                      k.GetAllParams(ctx),
-		FinalityProviders:           fps,
-		BtcDelegations:              dels,
-		BlockHeightChains:           k.blockHeightChains(ctx),
-		BtcDelegators:               btcDels,
-		Events:                      evts,
-		AllowedStakingTxHashes:      txHashes,
-		AllowedMultiStakingTxHashes: multiStakingTxHashes,
-		LargestBtcReorg:             k.GetLargestBtcReorg(ctx),
-		ConsumerEvents:              k.consumerEvents(ctx),
-		FpBbnAddr:                   fpBbnAddr,
-		DeletedFpsBtcPkHex:          deletedFps,
+		Params:             k.GetAllParams(ctx),
+		FinalityProviders:  fps,
+		BtcDelegations:     dels,
+		BlockHeightChains:  k.blockHeightChains(ctx),
+		BtcDelegators:      btcDels,
+		Events:             evts,
+		LargestBtcReorg:    k.GetLargestBtcReorg(ctx),
+		ConsumerEvents:     k.consumerEvents(ctx),
+		FpBbnAddr:          fpBbnAddr,
+		DeletedFpsBtcPkHex: deletedFps,
 	}, nil
 }
 
@@ -252,56 +217,6 @@ func (k Keeper) btcDelegatorsWithKey(ctx context.Context, storeKey []byte) ([]*t
 	}
 
 	return dels, nil
-}
-
-// allowedStakingTxHashes loads all allowed staking transactions hashes stored.
-// It encodes the hashes as hex strings to be human readable on exporting the genesis
-// This function has high resource consumption and should be only used on export genesis.
-func (k Keeper) allowedStakingTxHashes(ctx context.Context) ([]string, error) {
-	hashes := make([]string, 0)
-	iterator, err := k.AllowedStakingTxHashesKeySet.Iterate(ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer iterator.Close()
-
-	for ; iterator.Valid(); iterator.Next() {
-		key, err := iterator.Key()
-		if err != nil {
-			return nil, err
-		}
-
-		// encode hash as a hex string
-		hashStr := hex.EncodeToString(key)
-		hashes = append(hashes, hashStr)
-	}
-
-	return hashes, nil
-}
-
-// allowedMultiStakingTxHashes loads all allowed multi-staking transactions hashes stored.
-// It encodes the hashes as hex strings to be human readable on exporting the genesis
-// This function has high resource consumption and should be only used on export genesis.
-func (k Keeper) allowedMultiStakingTxHashes(ctx context.Context) ([]string, error) {
-	hashes := make([]string, 0)
-	iterator, err := k.allowedMultiStakingTxHashesKeySet.Iterate(ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer iterator.Close()
-
-	for ; iterator.Valid(); iterator.Next() {
-		key, err := iterator.Key()
-		if err != nil {
-			return nil, err
-		}
-
-		// encode hash as a hex string
-		hashStr := hex.EncodeToString(key)
-		hashes = append(hashes, hashStr)
-	}
-
-	return hashes, nil
 }
 
 // eventIdxs sets an event into the store.